#WSC-3_2: Products pagination, improvement, fixed ctx incorrect initialization

+++ internal/utils/pagination/paged_result.go
@@ -3,17 +3,17 @@ package pagination
 import "math"
 
 type PagedResult[T any] struct {
-	Data           []T  `json:"data"`
-	HasPrevPage    bool `json:"hasPrevPage"`
-	HasNextPage    bool `json:"hasNextPage"`
-	TotalCount     int  `json:"totalCount"`
-	Count          int  `json:"count"`
-	Page           int  `json:"page"`
-	TotalPageCount int  `json:"totalPageCount"`
+	Data           []T   `json:"data"`
+	HasPrevPage    bool  `json:"hasPrevPage"`
+	HasNextPage    bool  `json:"hasNextPage"`
+	TotalCount     int64 `json:"totalCount"`
+	Count          int   `json:"count"`
+	Page           int64 `json:"page"`
+	TotalPageCount int64 `json:"totalPageCount"`
 }
 
-func CreatePagedResult[T any](data []T, totalCount, page, pageSize int) PagedResult[T] {
-	totalPageCount := int(math.Ceil(float64(totalCount) / float64(pageSize)))
+func CreatePagedResult[T any](data []T, totalCount, page, pageSize int64) PagedResult[T] {
+	totalPageCount := int64(math.Ceil(float64(totalCount) / float64(pageSize)))
 	hasPrevPage := page > 1
 	hasNextPage := page < totalPageCount
